requirements.txt
# Core dependencies for Intraday Momentum OI Strategy Backtesting

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Backtesting framework
backtrader>=1.9.78.123

# Configuration
PyYAML>=6.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and utilities
python-dateutil>=2.8.0

# Optional: For enhanced analytics
scipy>=1.10.0
